Split index/repack functions to different files
